+++ src/box.cuh
@@ -23,15 +23,15 @@ public:
     void read_file(char*, int);
     ~Box(void);
     int *triclinic;
-    double* h; // GPU box data
+    float* h; // GPU box data
     // energy for the whole box 
-    double *pe_ref, *sxx_ref, *syy_ref, *szz_ref;
-    double *cpu_pe_ref, *cpu_sxx_ref, *cpu_syy_ref, *cpu_szz_ref;
-    double potential_square_sum;
-    double virial_square_sum;
+    float *pe_ref, *sxx_ref, *syy_ref, *szz_ref;
+    float *cpu_pe_ref, *cpu_sxx_ref, *cpu_syy_ref, *cpu_szz_ref;
+    float potential_square_sum;
+    float virial_square_sum;
     // functions
-    double get_volume(int, double*); // get the volume of the box
-    void get_inverse(int, double*);  // get the inverse box matrix
+    float get_volume(int, float*); // get the volume of the box
+    void get_inverse(int, float*);  // get the inverse box matrix
 };
