+++ src/fitness.cu
@@ -136,24 +136,24 @@ void Fitness::read_Na(FILE* fid)
 void Fitness::read_xyz(FILE* fid)
 {
     int m1 = sizeof(int) * N;
-    int m2 = sizeof(double) * N;
+    int m2 = sizeof(float) * N;
 
     CHECK(cudaMallocManaged((void**)&type, m1));
     CHECK(cudaMallocManaged((void**)&x, m2));
     CHECK(cudaMallocManaged((void**)&y, m2));
     CHECK(cudaMallocManaged((void**)&z, m2));
 
-    MY_MALLOC(cpu_fx_ref, double, N);
-    MY_MALLOC(cpu_fy_ref, double, N);
-    MY_MALLOC(cpu_fz_ref, double, N);
-    MY_MALLOC(cpu_fx, double, N);
-    MY_MALLOC(cpu_fy, double, N);
-    MY_MALLOC(cpu_fz, double, N);
+    MY_MALLOC(cpu_fx_ref, float, N);
+    MY_MALLOC(cpu_fy_ref, float, N);
+    MY_MALLOC(cpu_fz_ref, float, N);
+    MY_MALLOC(cpu_fx, float, N);
+    MY_MALLOC(cpu_fy, float, N);
+    MY_MALLOC(cpu_fz, float, N);
     num_types = 0;
     force_square_sum = 0.0;
     for (int n = 0; n < N; n++)
     {
-        int count = fscanf(fid, "%d%lf%lf%lf%lf%lf%lf", 
+        int count = fscanf(fid, "%d%f%f%f%f%f%f", 
             &(type[n]), &(x[n]), &(y[n]), &(z[n]),
             &(cpu_fx_ref[n]), &(cpu_fy_ref[n]), &(cpu_fz_ref[n]));
         if (count != 7) { print_error("reading error for xyz.in.\n"); }
@@ -176,7 +176,7 @@ void Fitness::read_xyz(FILE* fid)
 
 void Fitness::allocate_memory_gpu(void)
 {
-    int m2 = sizeof(double) * N;
+    int m2 = sizeof(float) * N;
     // read from CPU
     CHECK(cudaMalloc((void**)&fx_ref, m2));
     CHECK(cudaMalloc((void**)&fy_ref, m2));
@@ -200,22 +200,22 @@ void Fitness::allocate_memory_gpu(void)
 void Fitness::compute
 (
     int population_size, int number_of_variables, 
-    double *parameters_min, double *parameters_max,
-    double* population, double* fitness
+    float *parameters_min, float *parameters_max,
+    float* population, float* fitness
 )
 {
-    double *parameters;
-    double *error_gpu, *error_cpu;
-    MY_MALLOC(error_cpu, double, Nc);
-    CHECK(cudaMalloc((void**)&error_gpu, sizeof(double) * Nc));
-    MY_MALLOC(parameters, double, number_of_variables);
+    float *parameters;
+    float *error_gpu, *error_cpu;
+    MY_MALLOC(error_cpu, float, Nc);
+    CHECK(cudaMalloc((void**)&error_gpu, sizeof(float) * Nc));
+    MY_MALLOC(parameters, float, number_of_variables);
     for (int n = 0; n < population_size; ++n)
     {
-        double* individual = population + n * number_of_variables;
+        float* individual = population + n * number_of_variables;
         for (int m = 0; m < number_of_variables; ++m)
         {
-            double a = parameters_min[m];
-            double b = parameters_max[m] - a;
+            float a = parameters_min[m];
+            float b = parameters_max[m] - a;
             parameters[m] = a + b * individual[m];
         }
         update_potential(parameters);
@@ -231,13 +231,13 @@ void Fitness::compute
 
 
 static void predict_energy_or_stress
-(FILE* fid, double* cpu_data, double* data, double* ref, int N, int Nc)
+(FILE* fid, float* cpu_data, float* data, float* ref, int N, int Nc)
 {
-    cudaMemcpy(cpu_data, data, sizeof(double)*N, cudaMemcpyDeviceToHost);
+    cudaMemcpy(cpu_data, data, sizeof(float)*N, cudaMemcpyDeviceToHost);
     for (int nc = NC_FORCE; nc < Nc; ++nc)
     {
         int offset = nc * MAX_ATOM_NUMBER;
-        double cpu_data_nc = 0.0;
+        float cpu_data_nc = 0.0;
         for (int m = 0; m < MAX_ATOM_NUMBER; ++m)
         {
             cpu_data_nc += cpu_data[offset + m];
@@ -249,28 +249,28 @@ static void predict_energy_or_stress
 
 void Fitness::predict
 (
-    char* input_dir, int number_of_variables, double *parameters_min, 
-    double *parameters_max, double* elite
+    char* input_dir, int number_of_variables, float *parameters_min, 
+    float *parameters_max, float* elite
 )
 {
-    double *parameters;
-    MY_MALLOC(parameters, double, number_of_variables);
+    float *parameters;
+    MY_MALLOC(parameters, float, number_of_variables);
     for (int m = 0; m < number_of_variables; ++m)
     {
-        double a = parameters_min[m];
-        double b = parameters_max[m] - a;
+        float a = parameters_min[m];
+        float b = parameters_max[m] - a;
         parameters[m] = a + b * elite[m];
     }
     update_potential(parameters);
     find_force();
     MY_FREE(parameters);
 
-    cudaMemcpy(cpu_fx, fx, sizeof(double)*N, cudaMemcpyDeviceToHost);
-    cudaMemcpy(cpu_fy, fy, sizeof(double)*N, cudaMemcpyDeviceToHost);
-    cudaMemcpy(cpu_fz, fz, sizeof(double)*N, cudaMemcpyDeviceToHost);
-    cudaMemcpy(cpu_fx_ref, fx_ref, sizeof(double)*N, cudaMemcpyDeviceToHost);
-    cudaMemcpy(cpu_fy_ref, fy_ref, sizeof(double)*N, cudaMemcpyDeviceToHost);
-    cudaMemcpy(cpu_fz_ref, fz_ref, sizeof(double)*N, cudaMemcpyDeviceToHost);
+    cudaMemcpy(cpu_fx, fx, sizeof(float)*N, cudaMemcpyDeviceToHost);
+    cudaMemcpy(cpu_fy, fy, sizeof(float)*N, cudaMemcpyDeviceToHost);
+    cudaMemcpy(cpu_fz, fz, sizeof(float)*N, cudaMemcpyDeviceToHost);
+    cudaMemcpy(cpu_fx_ref, fx_ref, sizeof(float)*N, cudaMemcpyDeviceToHost);
+    cudaMemcpy(cpu_fy_ref, fy_ref, sizeof(float)*N, cudaMemcpyDeviceToHost);
+    cudaMemcpy(cpu_fz_ref, fz_ref, sizeof(float)*N, cudaMemcpyDeviceToHost);
 
     char file_force[200];
     strcpy(file_force, input_dir);
@@ -288,7 +288,7 @@ void Fitness::predict
     strcpy(file, input_dir);
     strcat(file, "/prediction.out");
     FILE* fid_prediction = my_fopen(file, "w");
-    double *cpu_prediction; MY_MALLOC(cpu_prediction, double, N);
+    float *cpu_prediction; MY_MALLOC(cpu_prediction, float, N);
     // energy
     predict_energy_or_stress
     (fid_prediction, cpu_prediction, pe, box.cpu_pe_ref, N, Nc);
@@ -306,7 +306,7 @@ void Fitness::predict
 }
 
 
-static __device__ void warp_reduce(volatile double* s, int t) 
+static __device__ void warp_reduce(volatile float* s, int t) 
 {
     s[t] += s[t + 32]; s[t] += s[t + 16]; s[t] += s[t + 8];
     s[t] += s[t + 4];  s[t] += s[t + 2];  s[t] += s[t + 1];
@@ -315,22 +315,22 @@ static __device__ void warp_reduce(volatile double* s, int t)
 
 static __global__ void gpu_sum_force_error
 (
-    int N, double *g_fx, double *g_fy, double *g_fz, 
-    double *g_fx_ref, double *g_fy_ref, double *g_fz_ref, double *g_error
+    int N, float *g_fx, float *g_fy, float *g_fz, 
+    float *g_fx_ref, float *g_fy_ref, float *g_fz_ref, float *g_error
 )
 {
     int tid = threadIdx.x;
     int number_of_patches = (N - 1) / 512 + 1; 
-    __shared__ double s_error[512];
+    __shared__ float s_error[512];
     s_error[tid] = 0.0;
     for (int patch = 0; patch < number_of_patches; ++patch)
     {
         int n = tid + patch * 512;
         if (n < N)
         {
-            double dx = g_fx[n] - g_fx_ref[n];
-            double dy = g_fy[n] - g_fy_ref[n];
-            double dz = g_fz[n] - g_fz_ref[n];
+            float dx = g_fx[n] - g_fx_ref[n];
+            float dy = g_fy[n] - g_fy_ref[n];
+            float dz = g_fz[n] - g_fz_ref[n];
             s_error[tid] += dx*dx + dy*dy + dz*dz;
         }
     }
@@ -343,25 +343,25 @@ static __global__ void gpu_sum_force_error
 }
 
 
-double Fitness::get_fitness_force(double *error_cpu, double *error_gpu)
+float Fitness::get_fitness_force(float *error_cpu, float *error_gpu)
 {
     int M = NC_FORCE * MAX_ATOM_NUMBER;
     gpu_sum_force_error<<<1, 512>>>(M, fx, fy, fz, 
         fx_ref, fy_ref, fz_ref, error_gpu);
-    CHECK(cudaMemcpy(error_cpu, error_gpu, sizeof(double), 
+    CHECK(cudaMemcpy(error_cpu, error_gpu, sizeof(float), 
         cudaMemcpyDeviceToHost));
     return sqrt(error_cpu[0] / force_square_sum);
 }
 
 
 static __global__ void gpu_sum_pe_error
-(int *g_Na, int *g_Na_sum, double *g_pe, double *g_pe_ref, double *error_gpu)
+(int *g_Na, int *g_Na_sum, float *g_pe, float *g_pe_ref, float *error_gpu)
 {
     int tid = threadIdx.x;
     int bid = blockIdx.x;
     int Na = g_Na[bid];
     int offset = g_Na_sum[bid];
-    __shared__ double s_pe[64];
+    __shared__ float s_pe[64];
     s_pe[tid] = 0.0;
     if (tid < Na)
     {
@@ -372,18 +372,18 @@ static __global__ void gpu_sum_pe_error
     if (tid < 32) { warp_reduce(s_pe, tid); }
     if (tid == 0)
     {
-        double diff = s_pe[0] - g_pe_ref[bid];
+        float diff = s_pe[0] - g_pe_ref[bid];
         error_gpu[bid] = diff * diff;
     }
 }
 
 
-double Fitness::get_fitness_energy(double* error_cpu, double* error_gpu)
+float Fitness::get_fitness_energy(float* error_cpu, float* error_gpu)
 {
     gpu_sum_pe_error<<<Nc, 64>>>(Na, Na_sum, pe, box.pe_ref, error_gpu);
-    int mem = sizeof(double) * Nc;
+    int mem = sizeof(float) * Nc;
     CHECK(cudaMemcpy(error_cpu, error_gpu, mem, cudaMemcpyDeviceToHost));
-    double error_ave = 0.0;
+    float error_ave = 0.0;
     for (int n = NC_FORCE; n < Nc; ++n)
     {
         error_ave += error_cpu[n];
@@ -392,10 +392,10 @@ double Fitness::get_fitness_energy(double* error_cpu, double* error_gpu)
 }
 
 
-double Fitness::get_fitness_stress(double* error_cpu, double* error_gpu)
+float Fitness::get_fitness_stress(float* error_cpu, float* error_gpu)
 {
-    double error_ave = 0.0;
-    int mem = sizeof(double) * Nc;
+    float error_ave = 0.0;
+    int mem = sizeof(float) * Nc;
 
     gpu_sum_pe_error<<<Nc, 64>>>(Na, Na_sum, sxx, box.sxx_ref, error_gpu);
     CHECK(cudaMemcpy(error_cpu, error_gpu, mem, cudaMemcpyDeviceToHost));
